fix(cli): Don't downgrade deployment target on migrate
